messaging_app/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.6
PyJWT==2.9.0
python-environ==0.4.54
sqlparse==0.5.3
typing-extensions==4.12.2
